+++ src/main/java/org/codinjutsu/tools/jenkins/JenkinsAppSettings.java
@@ -20,9 +20,9 @@ import com.intellij.openapi.components.PersistentStateComponent;
 import com.intellij.openapi.components.State;
 import com.intellij.openapi.components.Storage;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.util.xmlb.XmlSerializerUtil;
 import lombok.Data;
-import org.apache.commons.lang.StringUtils;
 import org.codinjutsu.tools.jenkins.model.Build;
 import org.codinjutsu.tools.jenkins.model.BuildStatusEnum;
 import org.jetbrains.annotations.NotNull;
@@ -38,6 +38,7 @@ public class JenkinsAppSettings implements PersistentStateComponent<JenkinsAppSe
     public static final String DUMMY_JENKINS_SERVER_URL = "http://dummyjenkinsserver";
     public static final int DEFAULT_BUILD_DELAY = 0;
     public static final int DEFAULT_BUILD_RETRY = 0;
+    public static final int DEFAULT_JOBS_PER_BUILD = 0;
     public static final int RESET_PERIOD_VALUE = 0;
 
     private State myState = new State();
@@ -79,7 +80,7 @@ public class JenkinsAppSettings implements PersistentStateComponent<JenkinsAppSe
     @Deprecated(since = "0.13.18")
     public boolean isServerUrlSet() {
         final String serverUrl = myState.getServerUrl();
-        return StringUtils.isNotEmpty(serverUrl) && !DUMMY_JENKINS_SERVER_URL.equals(serverUrl);
+        return StringUtil.isNotEmpty(serverUrl) && !DUMMY_JENKINS_SERVER_URL.equals(serverUrl);
     }
 
     public int getBuildDelay() {
@@ -130,8 +131,8 @@ public class JenkinsAppSettings implements PersistentStateComponent<JenkinsAppSe
         return getRssSettings().isDisplayAborted();
     }
 
-    public void setIgnoreSuccessOrStable(boolean ignoreSucessOrStable) {
-        getRssSettings().setDisplaySuccessOrStable(ignoreSucessOrStable);
+    public void setDisplaySuccessOrStable(boolean displaySuccessOrStable) {
+        getRssSettings().setDisplaySuccessOrStable(displaySuccessOrStable);
     }
 
     public void setDisplayUnstableOrFail(boolean displayUnstableOrFail) {
@@ -206,6 +207,14 @@ public class JenkinsAppSettings implements PersistentStateComponent<JenkinsAppSe
         myState.setShowLogIfTriggerBuild(showLogIfTriggerBuild);
     }
 
+    public void setBuildsToLoadPerJob(int buildsToLoadPerJobNumber) {
+        myState.setBuildsToLoadPerJob(buildsToLoadPerJobNumber);
+    }
+
+    public int getBuildsToLoadPerJob() {
+        return myState.getBuildsToLoadPerJob();
+    }
+
     @Data
     public static class State {
 
@@ -216,6 +225,7 @@ public class JenkinsAppSettings implements PersistentStateComponent<JenkinsAppSe
         private String suffix = "";
 
         private int numBuildRetries = DEFAULT_BUILD_RETRY;
+        private int buildsToLoadPerJob = DEFAULT_JOBS_PER_BUILD;
         private RssSettings rssSettings = new RssSettings();
         private boolean useGreenColor = false;
         private boolean showAllInStatusbar = false;