+++ src/main/java/org/codinjutsu/tools/jenkins/logic/RequestManager.java
@@ -17,14 +17,15 @@
 package org.codinjutsu.tools.jenkins.logic;
 
 import com.intellij.openapi.Disposable;
+import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.offbytwo.jenkins.JenkinsServer;
 import com.offbytwo.jenkins.helper.BuildConsoleStreamListener;
 import com.offbytwo.jenkins.model.*;
-import org.apache.commons.lang.StringUtils;
 import org.codinjutsu.tools.jenkins.JenkinsAppSettings;
 import org.codinjutsu.tools.jenkins.JenkinsSettings;
 import org.codinjutsu.tools.jenkins.exception.*;
@@ -89,7 +90,7 @@ public class RequestManager implements RequestManagerInterface, Disposable {
         var jenkinsWorkspaceData = securityClient.execute(url);
         final var serverUrl = configuration.getServerUrl();
         final var configuredJenkinsUrl = Optional.of(jenkinsSettings.getJenkinsUrl())
-                .filter(StringUtils::isNotEmpty)
+                .filter(StringUtil::isNotEmpty)
                 .orElse(serverUrl);
 
         if (configuredJenkinsUrl.contains(BUILDHIVE_CLOUDBEES)) {//TODO hack need to refactor
@@ -250,9 +251,9 @@ public class RequestManager implements RequestManagerInterface, Disposable {
         return jenkinsParser.createBuild(jenkinsJobData);
     }
 
-    private List<Build> loadBuilds(String jenkinsBuildUrl) {
+    private List<Build> loadBuilds(String jenkinsBuildUrl, RangeToLoad rangeToLoad) {
         if (handleNotYetLoggedInState()) return Collections.emptyList();
-        URL url = urlBuilder.createBuildsUrl(jenkinsBuildUrl);
+        URL url = urlBuilder.createBuildsUrl(jenkinsBuildUrl, rangeToLoad);
         String jenkinsJobData = securityClient.execute(url);
         return jenkinsParser.createBuilds(jenkinsJobData);
     }
@@ -300,7 +301,8 @@ public class RequestManager implements RequestManagerInterface, Disposable {
         securityClient.connect(urlBuilder.createAuthenticationUrl(serverUrl));
         setJenkinsServer(new JenkinsServer(new JenkinsClient(urlBuilder.createServerUrl(serverUrl), securityClient)));
 
-        final var urlMapper = project.getService(UrlMapperService.class).getMapper(jenkinsSettings, serverUrl);;
+        final var urlMapper = ApplicationManager.getApplication().getService(UrlMapperService.class)
+                .getMapper(jenkinsSettings, serverUrl);;
         setJenkinsParser(new JenkinsJsonParser(urlMapper));
     }
 
@@ -311,7 +313,7 @@ public class RequestManager implements RequestManagerInterface, Disposable {
         SecurityClientFactory.setVersion(version);
         final int connectionTimout = getConnectionTimout(connectionTimoutInSeconds);
         final SecurityClient securityClientForTest;
-        if (StringUtils.isNotBlank(username)) {
+        if (org.codinjutsu.tools.jenkins.util.StringUtil.isNotBlank(username)) {
             securityClientForTest = SecurityClientFactory.basic(username, password, crumbData, connectionTimout);
         } else {
             securityClientForTest = SecurityClientFactory.none(crumbData, connectionTimout);
@@ -348,7 +350,8 @@ public class RequestManager implements RequestManagerInterface, Disposable {
 
     @Override
     public List<Build> loadBuilds(Job job) {
-        return loadBuilds(job.getUrl());
+        final var rangeToLoad = RangeToLoad.to(JenkinsAppSettings.getSafeInstance(project).getBuildsToLoadPerJob());
+        return loadBuilds(job.getUrl(), rangeToLoad);
     }
 
     @NotNull