+++ src/main/java/org/codinjutsu/tools/jenkins/logic/UrlBuilder.java
@@ -89,9 +89,12 @@ public class UrlBuilder {
                 LOG.debug(e.getMessage(), e);
             }
             //return com.intellij.util.io.URLUtil.encodeQuery(query);
+            // prevent double encode of ',={}'
             return path + '?' + URLUtil.encodeURIComponent(query)
                     .replaceAll("%3D", "=")
-                    .replaceAll("%2C", ",");
+                    .replaceAll("%2C", ",")
+                    .replaceAll("%257B", "%7B")
+                    .replaceAll("%257D", "%7D");
         }
     }
 
@@ -131,8 +134,8 @@ public class UrlBuilder {
         return buildUrl(buildUrl, encodePathQuery(API_JSON + TREE_PARAM + BASIC_BUILD_INFO));
     }
 
-    public URL createBuildsUrl(String buildUrl) {
-        return buildUrl(buildUrl, encodePathQuery(API_JSON + TREE_PARAM + BASIC_BUILDS_INFO));
+    public URL createBuildsUrl(String buildUrl, RangeToLoad rangeToLoad) {
+        return buildUrl(buildUrl, encodePathQuery(API_JSON + TREE_PARAM + BASIC_BUILDS_INFO + rangeToLoad.toQueryParameter()));
     }
 
     public URL createRssLatestUrl(String serverUrl) {
@@ -152,7 +155,7 @@ public class UrlBuilder {
         return null;
     }
 
-    private void handleException(Exception ex) {
+    private static void handleException(Exception ex) {
         LOG.debug(ex);
         if (ex instanceof MalformedURLException) {
             throw new IllegalArgumentException("URL is malformed", ex);
@@ -187,6 +190,17 @@ public class UrlBuilder {
         }
     }
 
+    @NotNull
+    public static URL createViewUrl(@NotNull String serverUrl, @NotNull String viewName) {
+        try {
+            return buildUrlNotNull(removeTrailingSlash(serverUrl), "/view/" +
+                    URLUtil.encodeURIComponent(viewName));
+        } catch (Exception ex) {
+            handleException(ex);
+            throw new IllegalArgumentException(ERROR_DURING_URL_CREATION, ex);
+        }
+    }
+
     @Nullable
     public URL toUrl(@NotNull String url) {
         try {
@@ -198,7 +212,7 @@ public class UrlBuilder {
     }
 
     @NotNull
-    public String removeTrailingSlash(@NotNull String url) {
+    public static String removeTrailingSlash(@NotNull String url) {
         final String withoutTrailingSlash;
         if (url.endsWith("/")) {
             withoutTrailingSlash = url.substring(0, url.length() - 1);
@@ -221,7 +235,7 @@ public class UrlBuilder {
         return null;
     }
 
-    private @NotNull URL buildUrlNotNull(String context, @NotNull String pathWithQuery) throws MalformedURLException {
+    private static @NotNull URL buildUrlNotNull(String context, @NotNull String pathWithQuery) throws MalformedURLException {
         final boolean pathWithLeadingSlash = StringUtil.startsWithChar(pathWithQuery, '/');
         final String serverContext = pathWithLeadingSlash ? UriUtil.trimTrailingSlashes(context) : context;
         return new URL(serverContext + pathWithQuery);