+++ src/main/java/org/codinjutsu/tools/jenkins/view/BrowserPanel.java
@@ -23,10 +23,10 @@ import com.intellij.openapi.components.*;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.SimpleToolWindowPanel;
+import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.ui.PopupHandler;
 import com.intellij.ui.ScrollPaneFactory;
 import com.intellij.util.ui.tree.TreeUtil;
-import org.apache.commons.lang.StringUtils;
 import org.codinjutsu.tools.jenkins.*;
 import org.codinjutsu.tools.jenkins.logic.*;
 import org.codinjutsu.tools.jenkins.model.*;
@@ -55,6 +55,7 @@ import java.util.stream.Collectors;
 @State(name = "JenkinsBrowserPanel", storages = {
         @Storage(value = StoragePathMacros.PRODUCT_WORKSPACE_FILE, roamingType = RoamingType.DISABLED)
 })
+@Service(Service.Level.PROJECT)
 public final class BrowserPanel extends SimpleToolWindowPanel implements PersistentStateComponent<JenkinsTreeState> {
 
     @NonNls
@@ -236,7 +237,7 @@ public final class BrowserPanel extends SimpleToolWindowPanel implements Persist
         return jenkins;
     }
 
-    public View getCurrentSelectedView() {
+    public @Nullable View getCurrentSelectedView() {
         return currentSelectedView;
     }
 
@@ -282,7 +283,7 @@ public final class BrowserPanel extends SimpleToolWindowPanel implements Persist
         GuiUtil.runInSwingThread(() -> {
             jobTree.updateJobNode(job);
             CollectionUtil.flattenedJobs(jenkins.getJobs()).forEach(j -> visit(j, buildStatusAggregator));
-            JenkinsWidget.getInstance(project).updateStatusIcon(buildStatusAggregator);
+            JenkinsStatusBarWidget.getInstance(project).updateStatusIcon(buildStatusAggregator);
         });
     }
 
@@ -314,7 +315,7 @@ public final class BrowserPanel extends SimpleToolWindowPanel implements Persist
 
     private void clearView() {
         jobTree.clear();
-        JenkinsWidget.getInstance(project).updateStatusIcon(BuildStatusAggregator.EMPTY);
+        JenkinsStatusBarWidget.getInstance(project).updateStatusIcon(BuildStatusAggregator.EMPTY);
     }
 
     public void postAuthenticationInitialization() {
@@ -324,7 +325,7 @@ public final class BrowserPanel extends SimpleToolWindowPanel implements Persist
 
         String lastSelectedViewName = jenkinsSettings.getLastSelectedView();
         View viewToLoad;
-        if (StringUtils.isEmpty(lastSelectedViewName)) {
+        if (StringUtil.isEmpty(lastSelectedViewName)) {
             viewToLoad = jenkins.getPrimaryView();
         } else if (favoriteView != null && lastSelectedViewName.equals(favoriteView.getName())) {
             viewToLoad = favoriteView;
@@ -543,7 +544,7 @@ public final class BrowserPanel extends SimpleToolWindowPanel implements Persist
 
             GuiUtil.runInSwingThread(() -> {
                 fillJobTree(buildStatusAggregator);
-                JenkinsWidget.getInstance(project).updateStatusIcon(buildStatusAggregator);
+                JenkinsStatusBarWidget.getInstance(project).updateStatusIcon(buildStatusAggregator);
             });
         }