+++ src/main/java/org/codinjutsu/tools/jenkins/view/JenkinsViewComboRenderer.java
@@ -16,23 +16,49 @@
 
 package org.codinjutsu.tools.jenkins.view;
 
-import com.intellij.ui.ColoredListCellRenderer;
-import com.intellij.ui.SimpleTextAttributes;
+import com.intellij.icons.AllIcons;
+import com.intellij.ui.JBColor;
+import com.intellij.ui.SimpleListCellRenderer;
+import com.intellij.ui.list.ListCellBackgroundSupplier;
+import com.intellij.util.ui.UIUtil;
 import org.codinjutsu.tools.jenkins.model.FavoriteView;
 import org.codinjutsu.tools.jenkins.model.View;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
+import java.awt.*;
 
-public class JenkinsViewComboRenderer extends ColoredListCellRenderer<View> {
+public class JenkinsViewComboRenderer extends SimpleListCellRenderer<View>
+        implements ListCellBackgroundSupplier<View> {
 
-    @Override
-    protected void customizeCellRenderer(@NotNull JList<? extends View> list, View view, int index, boolean selected, boolean hasFocus) {
-        append(view.getName(), SimpleTextAttributes.REGULAR_ATTRIBUTES);
-
+    public void renderView(View view) {
+        final boolean isNestedParent = view.hasNestedView();
+        setEnabled(!isNestedParent);
+        setFocusable(!isNestedParent);
+        if (isNestedParent) {
+            setIcon(AllIcons.Nodes.Folder);
+            setBackground(getCellBackground(view, 0));
+            setForeground(UIUtil.getLabelDisabledForeground());
+        }
+        setText(getText(view));
         if (view instanceof FavoriteView) {
             setIcon(JenkinsTreeRenderer.FAVORITE_ICON);
         }
+    }
 
+    @Override
+    public void customize(@NotNull JList<? extends View> list, View view, int index, boolean selected, boolean hasFocus) {
+        renderView(view);
+    }
+
+    @Override
+    public @Nullable Color getCellBackground(View view, int row) {
+        return view.hasNestedView() ? JBColor.LIGHT_GRAY : null;
+    }
+
+    private @NotNull String getText(View view) {
+        final var viewName = view.getName();
+        return view.isNested() ? "   " + viewName : viewName;
     }
 }