+++ src/main/java/org/codinjutsu/tools/jenkins/view/action/settings/SortByStatusAction.java
@@ -17,10 +17,12 @@
 package org.codinjutsu.tools.jenkins.view.action.settings;
 
 import com.intellij.icons.AllIcons;
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.ToggleAction;
 import com.intellij.openapi.project.DumbAware;
 import org.codinjutsu.tools.jenkins.view.BrowserPanel;
+import org.jetbrains.annotations.NotNull;
 
 import javax.swing.*;
 
@@ -46,6 +48,11 @@ public class SortByStatusAction extends ToggleAction implements DumbAware {
     public void setSelected(AnActionEvent anActionEvent, boolean sorted) {
         sortedByStatus = sorted;
         browserPanel.setSortedByStatus(sorted);
+    }
+
+    @Override
+    public @NotNull ActionUpdateThread getActionUpdateThread() {
+        return ActionUpdateThread.BGT;
     }
 
 }